chore: export function again
